requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-crispy-forms==1.7.2
django-filter==2.2.0
djangorestframework==3.10.3
httpie==1.0.3
idna==2.8
Markdown==3.1.1
Pillow==6.2.0
pkg-resources==0.0.0
pycodestyle==2.5.0
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
requests-cache==0.5.2
sqlparse==0.3.0
urllib3==1.25.6
